+++ src/components/common/Header/index.tsx
@@ -1,10 +1,35 @@
 import Logo from 'assets/logo.svg';
 import { useNavigate } from 'react-router-dom';
+import { deleteAuthority, deleteSchoolName } from 'lib/storage/user';
+import { deleteAccessToken, deleteRefreshToken } from 'lib/storage/token';
+import { useRecoilState } from 'recoil';
+import { useMutation } from 'react-query';
+import { userData } from 'atom/user';
+import { logoutUser } from 'api/auth';
 import SearchInput from '../SearchInput';
+
 import * as S from './style';
 
 const Header = () => {
   const navigate = useNavigate();
+  const [user, setUser] = useRecoilState(userData);
+  const logoutMutate = useMutation(logoutUser, {
+    onSuccess: () => {
+      deleteAuthority();
+      deleteSchoolName();
+      deleteAccessToken();
+      deleteRefreshToken();
+      alert('로그아웃 성공 !!');
+      window.location.reload();
+    },
+    onError: (err) => {
+      console.log(err);
+    },
+  });
+
+  const logout = () => {
+    logoutMutate.mutate();
+  };
   return (
     <S.Container>
       <S.Wrap>
@@ -20,7 +45,11 @@ const Header = () => {
             type="text"
             name="stupetitionSearch"
           />
-          <S.Login onClick={() => navigate('/login')}>로그인</S.Login>
+          {user?.authority ? (
+            <S.Logout onClick={logout}>로그아웃</S.Logout>
+          ) : (
+            <S.Login onClick={() => navigate('/login')}>로그인</S.Login>
+          )}
         </S.NavWrap>
       </S.Wrap>
     </S.Container>
